fix slug page
