--- components/header.tsx
@@ -1,9 +1,0 @@
-import React from 'react'
-// import Link from 'next/link'
-import { Heading } from 'rebass'
-
-const Header: React.FC = () => {
-  return <Heading color="primary">Shoota</Heading>
-}
-
-export default Header
