--- components/post-header.tsx
@@ -1,39 +1,0 @@
-import React from 'react'
-
-import Author from '../types/author'
-
-import Avatar from './atoms/Avatar'
-import DateFormatter from './atoms/DateFormatter'
-import { CoverImage } from './CoverImage'
-import PostTitle from './post-title'
-
-type Props = {
-  title: string
-  coverImage: string
-  date: string
-  author: Author
-}
-
-const PostHeader: React.FC<Props> = ({ title, coverImage, date, author }) => {
-  return (
-    <>
-      <PostTitle>{title}</PostTitle>
-      <div className="hidden md:block md:mb-12">
-        <Avatar name={author.name} picture={author.picture} />
-      </div>
-      <div className="mb-8 md:mb-16 sm:mx-0">
-        <CoverImage title={title} src={coverImage} />
-      </div>
-      <div className="max-w-2xl mx-auto">
-        <div className="block md:hidden mb-6">
-          <Avatar name={author.name} picture={author.picture} />
-        </div>
-        <div className="mb-6 text-lg">
-          <DateFormatter dateString={date} />
-        </div>
-      </div>
-    </>
-  )
-}
-
-export default PostHeader
