+++ pages/posts/[slug].tsx
@@ -5,11 +5,10 @@ import Head from 'next/head'
 
 import { ContainerBox } from '../../components/ContainerBox'
 import PostBody from '../../components/post-body'
-import Header from '../../components/header'
-import PostHeader from '../../components/post-header'
+import PostHeader from '../../components/PostHeader'
 import Layout from '../../components/layout'
 import { getPostBySlug, getAllPosts } from '../../lib/api'
-import PostTitle from '../../components/post-title'
+import { PostTitle } from '../../components/PostTitle'
 import { CMS_NAME } from '../../lib/constants'
 import markdownToHtml from '../../lib/markdownToHtml'
 import PostType from '../../types/post'
@@ -28,12 +27,11 @@ const Post: React.FC<Props> = ({ post }) => {
   return (
     <Layout>
       <ContainerBox>
-        <Header />
         {router.isFallback ? (
           <PostTitle>Loading…</PostTitle>
         ) : (
           <>
-            <article className="mb-32">
+            <article>
               <Head>
                 <title>
                   {post.title} | Next.js Blog Example with {CMS_NAME}
